Extract WithStandardTopMargin out
