+++ app/client/components/MMAPageSkeleton.tsx
@@ -6,8 +6,9 @@ import {
   shouldHaveHolidayStopsFlow
 } from "../../shared/productTypes";
 import { MenuSpecificNavItem, NAV_LINKS } from "./nav/navConfig";
-import { PageContainer, WithStandardTopMargin } from "./page";
+import { PageContainer } from "./page";
 import { Spinner } from "./spinner";
+import { WithStandardTopMargin } from "./WithStandardTopMargin";
 
 interface LocationObject {
   title: string;
