+++ app/client/components/cancel/cancellationFlow.tsx
@@ -17,9 +17,9 @@ import { maxWidth } from "../../styles/breakpoints";
 import { LinkButton } from "../buttons";
 import { FlowWrapper } from "../FlowWrapper";
 import { NAV_LINKS } from "../nav/navConfig";
-import { WithStandardTopMargin } from "../page";
 import { ProgressIndicator } from "../progressIndicator";
 import { RadioButton } from "../radioButton";
+import { WithStandardTopMargin } from "../WithStandardTopMargin";
 import {
   ReturnToAccountOverviewButton,
   RouteableStepProps,
