+++ app/client/components/cancel/gw/gwCancellationFlowStart.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { trackEvent } from "../../analytics";
-import { WithStandardTopMargin } from "../../page";
+import { WithStandardTopMargin } from "../../WithStandardTopMargin";
 import { hrefStyle } from "../cancellationConstants";
 
 const trackCancellationClickEvent = (eventLabel: string) => () =>
