+++ app/client/components/identity/Settings/index.tsx
@@ -2,8 +2,9 @@ import * as Sentry from "@sentry/browser";
 import React, { useEffect, useState } from "react";
 import { trackEvent } from "../../analytics";
 import { NAV_LINKS } from "../../nav/navConfig";
-import { PageContainer, WithStandardTopMargin } from "../../page";
+import { PageContainer } from "../../page";
 import { Spinner } from "../../spinner";
+import { WithStandardTopMargin } from "../../WithStandardTopMargin";
 import {
   GenericErrorMessage,
   GenericErrorMessageRef
